+++ lib/type-functions/base.ts
@@ -11,12 +11,16 @@ export function getBaseType(input: {
   definition: Definition;
   name: string;
   scope: 'create' | 'update' | 'output';
+  skipFields?: string[];
 }) {
   const decoratorFn = input.scope === 'output' ? ObjectType : InputType;
   @decoratorFn(input.name)
   class Placeholder {}
 
   for (const property of inspect(input.definition).getProperties()) {
+    if (input.skipFields?.includes(property.name)) {
+      continue;
+    }
     const designType = Reflect.getMetadata('design:type', input.definition.prototype, property.name);
     if (designType === ObjectId || property.get(MetaKey.Property)?.db?.type?.[0] === ObjectId) {
       Transform(({ obj, key }) => obj[key])(Placeholder.prototype, property.name);
