+++ src/models/index.ts
@@ -1,6 +1,5 @@
 export * from './author';
 export * from './user';
 export * from './product';
-export * from './variant';
 export * from './customer';
 export * from './computer';
